+++ kosogor/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+# 1.0.3 - 2019-03-20
+- Add task getter to ShadowJarProxy
+- Add possibility to override main jar for PublishJarProxy
+
 # 1.0.2 - 2019-03-13
 - Fix BuildDir defaults broken in 1.0.2 (it was not applied to parent projects)
 - Update wrapper defaults to 5.2.1